+++ package.json
@@ -30,6 +30,7 @@
     "nodemon": "^2.0.7",
     "path": "^0.12.7",
     "ramda": "^0.27.1",
+    "sinon": "^15.0.1",
     "yup": "^0.32.9"
   },
   "devDependencies": {
@@ -56,5 +57,8 @@
     "src/**/**/*.js": [
       "npm run lint"
     ]
+  },
+  "ava": {
+    "concurrency": 1
   }
 }
